marketplace-data-collector-main/python_service/requirements.txt
pandas==2.2.3
requests==2.32.3
selenium==4.30.0
psycopg2-binary
python-telegram-bot
undetected-chromedriver
fastapi[standard]
APScheduler==3.10.4
tenacity==8.2.3
attrs==25.1.0
cachetools==5.5.2
certifi==2025.1.31
charset-normalizer==3.4.1
et_xmlfile==2.0.0
google-auth==2.40.1
google-auth-oauthlib==1.2.2
gspread==6.2.0
h11==0.14.0
httplib2==0.22.0
idna==3.10
numpy==2.2.3
oauth2client==4.1.3
oauthlib==3.2.2
openpyxl==3.1.5
outcome==1.3.0.post0
packaging==24.2
pyasn1==0.6.1
pyasn1_modules==0.4.2
pyparsing==3.2.3
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2025.1
requests-oauthlib==2.0.0
rsa==4.9.1
selenium-stealth==1.0.6
six==1.17.0
sniffio==1.3.1
sortedcontainers==2.4.0
trio==0.29.0
trio-websocket==0.12.2
typing_extensions==4.12.2
tzdata==2025.1
urllib3==2.3.0
webdriver-manager==4.0.2
websocket-client==1.8.0
wsproto==1.2.0
